fix(layout): adjust app layout spacing to prevent navbar overlap
