+++ firebase-data/auth_export/config.json
@@ -1,1 +1,4 @@
-{"signIn":{"allowDuplicateEmails":false},"emailPrivacyConfig":{"enableImprovedEmailPrivacy":false}}
+{
+  "signIn": { "allowDuplicateEmails": false },
+  "emailPrivacyConfig": { "enableImprovedEmailPrivacy": false }
+}
