+++ src/app/opportunities/components/OpportunityDeleteDialog.tsx
@@ -1,0 +1,66 @@
+'use client';
+
+import { Warning as WarningIcon } from '@mui/icons-material';
+import {
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogContentText,
+  DialogActions,
+  Button,
+  CircularProgress,
+} from '@mui/material';
+import { useState } from 'react';
+
+interface OpportunityDeleteDialogProps {
+  open: boolean;
+  opportunity?: { id?: string; name?: string };
+  onCancel: () => void;
+  onConfirm: () => void;
+}
+
+export default function OpportunityDeleteDialog({
+  open,
+  opportunity,
+  onCancel,
+  onConfirm,
+}: OpportunityDeleteDialogProps) {
+  const [isDeleting, setIsDeleting] = useState(false);
+
+  const handleConfirm = async () => {
+    setIsDeleting(true);
+    try {
+      await onConfirm();
+    } finally {
+      setIsDeleting(false);
+    }
+  };
+
+  return (
+    <Dialog open={open} onClose={onCancel} maxWidth="xs" fullWidth>
+      <DialogTitle sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+        <WarningIcon color="error" />
+        Confirm Deletion
+      </DialogTitle>
+      <DialogContent>
+        <DialogContentText>
+          Are you sure you want to delete <strong>{opportunity?.name}</strong>? This
+          action cannot be undone.
+        </DialogContentText>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onCancel} variant="outlined">
+          Cancel
+        </Button>
+        <Button
+          onClick={handleConfirm}
+          color="error"
+          variant="contained"
+          disabled={isDeleting}
+        >
+          {isDeleting ? <CircularProgress size={20} /> : 'Delete'}
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
